start read_line with an 8-byte buffer to reduce reallocations
The shortest common message from Redis should be "OK\r\n", which is four bytes.
It is not uncommon for Vec-like data structures (including Vec) to double in size
every time they need to reallocate. Being off by a factor of up to 2 should be an
acceptable margin, especially since the "OK\r\n" message itself gets converted to
an enum value shortly after this, and that allocation is destroyed. Most other
uses of read_line are likely to use more than 8 bytes, but this allows us to avoid
the first three reallocations, at least.